[KED-2506] Ensure cli.py is optional (#1158)
