+++ tests/framework/cli/conftest.py
@@ -32,6 +32,7 @@ this directory. You don't need to import the fixtures as pytest will
 discover them automatically. More info here:
 https://docs.pytest.org/en/latest/fixture.html
 """
+import shutil
 import sys
 import tempfile
 from importlib import import_module
@@ -72,8 +73,11 @@ def entry_point(mocker, entry_points):
 @fixture(scope="module")
 def fake_root_dir():
     # using tempfile as tmp_path fixture doesn't support module scope
-    with tempfile.TemporaryDirectory() as tmp_root:
-        yield Path(tmp_root).resolve()
+    tmpdir = tempfile.mkdtemp()
+    try:
+        yield Path(tmpdir).resolve()
+    finally:
+        shutil.rmtree(tmpdir, ignore_errors=True)
 
 
 @fixture(scope="module")
